requirements.txt
autobahn==18.7.1
certifi==2019.11.28
cryptography==2.3
dateparser==0.7.0
pyOpenSSL==19.0.0
requests==2.22.0
service-identity==17.0.0
Twisted==19.7.0
ujson
python-binance==0.7.10
python-dotenv
numpy